fix: overlays and window
