+++ ArabicaCliento.sln
@@ -9,6 +9,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Content.Client", "space-sta
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Robust.Client", "space-station-14\RobustToolbox\Robust.Client\Robust.Client.csproj", "{AF33C396-4535-4816-92EA-1357BFBA4B71}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Content.Shared", "space-station-14\Content.Shared\Content.Shared.csproj", "{9C568A0F-208C-491A-9226-5E149B361304}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Robust.Shared", "space-station-14\RobustToolbox\Robust.Shared\Robust.Shared.csproj", "{ECD32E31-6A31-4AC9-B8A8-8904A9997515}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -30,5 +34,13 @@ Global
 		{AF33C396-4535-4816-92EA-1357BFBA4B71}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AF33C396-4535-4816-92EA-1357BFBA4B71}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AF33C396-4535-4816-92EA-1357BFBA4B71}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9C568A0F-208C-491A-9226-5E149B361304}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9C568A0F-208C-491A-9226-5E149B361304}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9C568A0F-208C-491A-9226-5E149B361304}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9C568A0F-208C-491A-9226-5E149B361304}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ECD32E31-6A31-4AC9-B8A8-8904A9997515}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ECD32E31-6A31-4AC9-B8A8-8904A9997515}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ECD32E31-6A31-4AC9-B8A8-8904A9997515}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ECD32E31-6A31-4AC9-B8A8-8904A9997515}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
