+++ ArabicaCliento/Commands/Arabica.ToggleCheatMenu.cs
@@ -1,5 +1,7 @@
 using ArabicaCliento.Systems;
+using ArabicaCliento.UI;
 using Content.Shared.Administration;
+using Robust.Client.UserInterface;
 using Robust.Shared.Console;
 
 namespace ArabicaCliento.Commands;
@@ -9,11 +11,11 @@ public class ToggleCheatMenuCommand: IConsoleCommand
 {
     public string Command => "arabica.toggle_menu";
     public string Description => "This command toggle the cheat menu";
+    
     public string Help => "arabica.toggle_menu";
     
     public void Execute(IConsoleShell shell, string argStr, string[] args)
     {
-        var sys = IoCManager.Resolve<IEntityManager>().System<ArabicaCheatMenuSystem>();
-        sys.ToggleMenu();
+        IoCManager.Resolve<IUserInterfaceManager>().GetUIController<ArabicaCheatMenuUiController>().ToggleMenu();
     }
 }
